fix calorie datatype

+++ DDUBUCK/app/src/main/java/com/mapo/ddubuck/data/mypagechart/WeekStat.kt
@@ -2,7 +2,7 @@ package com.mapo.ddubuck.data.mypagechart
 
 data class WeekStat(
     val name: String,
-    val calorie: Int,
+    val calorie: Double,
     val completedCount: Int,
     val created_at: String,
     val stepCount: Int,
